Reimplement --deps-only
Working around a set of bugs that were in the original version, but also
renouncing using the `--formula` code path because that was too limited for
using deps-only on multiple packages, and with unspecified versions.

This creates virtual packages at the client level that mimic the dependencies of
the specified packages (except for dependencies between them). This resolves a
few issues:

- these packages are always available (the original implem patched the packages
  but that could result in unavailable packages getting installed) ; they are
  however marked as avoid-version if unavailable as a cheap way to stay on the
  right version when possible.

- they exist in parallel to their "real" counterpart, so that these can still be
  installed if required through dependency chains, keeping consistency of the
  solution.

- they forbid different versions of their "real" counterpart, to ensure
  consistency.

- the `deps-of-xxx` package are removed from the solution, but could appear in
  conflict messages. I'd argue that it should make things understandable even if
  it is an implementation detail.

+++ src/solver/opamSolver.ml
@@ -729,7 +729,8 @@ let string_of_stats stats =
 let solution_is_empty t =
   OpamCudf.ActionGraph.is_empty t
 
-let print_solution ~messages ~append ~requested ~reinstall ~available t =
+let print_solution ~messages ~append ~requested ~reinstall ~available
+    ?(skip=OpamPackage.Map.empty) t =
   let dump_cudf sfx t = match OpamSolverConfig.(!r.cudf_file) with
     | None -> ()
     | Some f ->
@@ -746,11 +747,18 @@ let print_solution ~messages ~append ~requested ~reinstall ~available t =
   in
   let actions, details =
     OpamCudf.ActionGraph.fold_vertex (fun a (actions,details) ->
+        if OpamPackage.Map.mem OpamCudf.(cudf2opam (action_contents a)) skip
+        then actions, details
+        else
         let cause =
           try OpamCudf.Map.find (OpamCudf.action_contents a) causes
           with Not_found -> Unknown in
         let action = map_action OpamCudf.cudf2opam a in
-        let cudf_name p = OpamPackage.name_to_string (OpamCudf.cudf2opam p) in
+        let cudf_name p =
+          let nv = OpamCudf.cudf2opam p in
+          let nv = try OpamPackage.Map.find nv skip with Not_found -> nv in
+          OpamPackage.name_to_string nv
+        in
         let cause = string_of_cause cudf_name cause in
         let messages =
           match a with
@@ -810,11 +818,11 @@ let dump_universe universe oc =
           (OpamPackage.name_to_string nv) i (OpamPackage.version_to_string nv)
     ) version_map
 
-let filter_solution filter t =
+let filter_solution ?(recursive=true) filter t =
   let t = OpamCudf.ActionGraph.copy t in
   let rec rm iter_deps v =
     if OpamCudf.ActionGraph.mem_vertex t v then (
-      iter_deps (rm iter_deps) t v;
+      if recursive then iter_deps (rm iter_deps) t v;
       OpamCudf.ActionGraph.remove_vertex t v
     ) in
   OpamCudf.ActionGraph.iter_vertex
